Ready to smoke testing code
