+++ web/src/components/admin/employee/AdminEmployeeAndKidsTabs.vue
@@ -6,7 +6,8 @@ Navigate between employee admin pages using tabs
     <v-tabs>
       <v-tab to="/admin/employees/list">{{ $t('Сотрудники') }}</v-tab>
       <v-tab to="/admin/employees/kids">{{ $t('Дети') }}</v-tab>
-      <v-tab to="/admin/employees/import">{{ $t('Импорт из Excel') }}</v-tab>
+      <v-tab to="/admin/employees/import">{{ $t('Импорт сотрудников из Excel') }}</v-tab>
+      <v-tab to="/admin/employees/kids/import">{{ $t('Импорт детей из Excel') }}</v-tab>
     </v-tabs>
     <router-view></router-view>
   </div>
